move common helpers out of cloud
